% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic.R
\name{start_game}
\alias{start_game}
\title{Start Game}
\usage{
start_game(players.id.play = 1, players.n = 5)
}
\arguments{
\item{players.id.play}{Player's id}

\item{players.n}{Number of players}
}
\value{
plot
}
\description{
This function can start to paly game on R
}
